requirements-pip.txt
djangorestframework==3.11.0
gunicorn==20.0.4